+++ imagine/scenario/type/RepeatX.java
@@ -1,0 +1,99 @@
+package imagine.scenario.type;
+
+import imagine.scenario.Scenario;
+
+import java.awt.Graphics2D;
+
+/**
+ * Class to represent a {@code ScenarioType}
+ * that has as characteristic repeat itself
+ * on the x axis to fill the {@code GamePanel}.
+ * 
+ * @author Daniel O Sousa
+ * @author DaveJosef
+ */
+public class RepeatX implements ScenarioType {
+
+    /**
+     * Field that stores the name of
+     * this {@code ScenarioType}.
+     */
+    private String name = "RepeatX";
+
+    /**
+     * The {@code Scenario} that has
+     * this {@code ScenarioType}.
+     */
+    private Scenario scenario;
+
+    /**
+     * Constructs a new {@code RepeatX}
+     * {@code ScenarioType} that characterizes
+     * the {@code Scenario} passed via the
+     * {@code scenario} parameter.
+     * 
+     * @param scenario the scenario that
+     * has this type
+     */
+    public RepeatX(Scenario scenario) {
+        storeScenario(scenario);
+    }
+
+    /**
+     * Stores the {@code scenario} parameter
+     * in the {@code scenario} field.
+     * 
+     * @param scenario the {@code scenario}
+     * to be stored.
+     */
+    private void storeScenario(Scenario scenario) {
+        if(scenario == null) {
+            throw new IllegalArgumentException("scenario cannot be null");
+        }
+
+        this.scenario = scenario;
+    }
+
+    /**
+     * Returns the name of this
+     * {@code ScenarioType}.
+     * 
+     * @return the name of this
+     * {@code ScenarioType}
+     */
+    @Override
+    public String getName() {
+        return this.name;
+    }
+
+    /**
+     * Draws the {@code Scenario} that has this
+     * {@code ScenarioType} in a way that it repeats
+     * itself to fill the {@code GamePanel} on the
+     * x axis.
+     * 
+     * @param g2 a {@code Graphics2D} instance used
+     * for drawing the said {@code Scenario}
+     */
+    @Override
+    public void draw(Graphics2D g2) {
+        if(scenario.getGamePanelWidth() == 0 || scenario.getGamePanelHeight() == 0) {
+            return;
+        }
+
+        int drawingX =
+            scenario.getApparentX() <= 0 ?
+            scenario.getApparentX() % scenario.getWidth() :
+            scenario.getApparentX() % scenario.getWidth() - scenario.getWidth();
+        while(drawingX < scenario.getGamePanelWidth()) {
+            g2.drawImage (
+                scenario.getCurrentFrame().getImage(),
+                drawingX, scenario.getApparentY(),
+                scenario.getWidth(), scenario.getHeight(),
+                null
+            );
+            drawingX += scenario.getWidth();
+        }
+    }
+    
+}
